Added support for variables through the Context
The initial issue was that variables were described in the Arithmetic.pp
grammar but not supported in the Visitor. We are resolving this problem
by adding a dedicated case in the Visitor.

Variables are identified by a string identifier (see the grammar) and
their value is computed with a closure. Defining a variable is done
using the `addVariable` method:

```php
$visitor->addVariable('foo', function () { return rand(0, 10); });
```

The Visitor was also responsible of internally maintaining a Context
(holding constants, functions and variables definitions) and this was
breaking SR principle.

The notion of Context has been extracted to a dedicated class allowing
more flexibility: we can stick with the default provided context or
build our own and inject it in the visitor thanks to the `setContext`
method.

The context also allows us to reach a high level of flexibility with
variables. Their value might change based on external element so the
context gets passed to the variable closure to allow computing their
values:

```php
$context->addVariable('foo', function (Context $c) { return rand(0,
$c->getVariable('max')); });
```

Thanks @esion for his initial work on this feature.

Closes #29